update app/frontendフォルダ内のフォルダのリネーム
javascripts→js, stylesheets→scss, packs→entry
