+++ app/frontend/entry/articles/show.js
@@ -1,6 +1,6 @@
-import FavoriteButton from '../../javascripts/components/favorite_button'
-import CommentDeleteButton from '../../javascripts/components/comment_delete_button'
-// import CommentForm from '../../javascripts/components/comment_form'
+import FavoriteButton from '../../js/components/favorite_button'
+import CommentDeleteButton from '../../js/components/comment_delete_button'
+// import CommentForm from '../../js/components/comment_form'
 
 document.addEventListener('DOMContentLoaded', function () {
   new vue({
