+++ app/frontend/entry/drafts/index.js
@@ -1,4 +1,4 @@
-import DeleteDraftButton from "../../javascripts/components/delete_draft_button";
+import DeleteDraftButton from "../../js/components/delete_draft_button";
 
 document.addEventListener("DOMContentLoaded", function () {
   new vue({
