+++ app/frontend/entry/static_pages/about.js
