+++ app/frontend/entry/users/show.js
@@ -1,6 +1,6 @@
 import '../default'
 // import Vue from "vue/dist/vue.esm.js";
-import FollowButton from "../../javascripts/components/follow_button";
+import FollowButton from "../../js/components/follow_button";
 
 document.addEventListener("DOMContentLoaded", function () {
   new vue({
