+++ app/frontend/js/components/comment_delete_button.vue
@@ -1,40 +1,49 @@
 <template>
-  <button @click="deleteDraft" type="button" class="btn btn-outline-danger">削除</button>
+  <button :disabled="isToastRunning" @click="deleteComment" class="btn btn-outline-danger">削除</button>
 </template>
 
 <script>
 import Axios from "axios";
 import { csrfToken } from "@rails/ujs";
 import { async } from "q";
-import Toast from "../mixin/toast";
+import Toast from "../mixins/toast";
 
 Axios.defaults.headers.common["X-CSRF-TOKEN"] = csrfToken();
 export default {
   mixins: [Toast],
   props: {
-    articleId: Number
+    commentId: Number
+  },
+  data: function() {
+    return {};
+  },
+  computed: {
+    isToastRunning: function() {
+      return Swal.isTimerRunning();
+    }
   },
   methods: {
-    deleteDraft: function() {
-      const element = this.$el.parentNode.parentNode;
+    deleteComment: function() {
+      const element = this.$el.parentNode;
       const toast = this.toast;
-      Swal.fire({
+      const confirmation = Swal.fire({
         type: "warning",
-        title: "下書きを削除してもよろしいですか？",
-        text: "この操作は取り消せません",
+        title: "コメントを削除してもよろしいですか？",
+        text: "この操作は取り消しできません",
         showCancelButton: true,
         confirmButtonText: "削除する",
         cancelButtonText: "キャンセル"
       }).then(result => {
         if (result.value) {
-          Axios.delete(`/drafts/${this.articleId}`)
-            .then(response => {
+          Axios.delete(`/comments/${this.commentId}`)
+            .then(function(response) {
+              console.log(response.data);
               if (response.status === 200) {
                 element.remove();
                 toast(response.data.type, response.data.message);
               }
             })
-            .catch(error => {
+            .catch(function(error) {
               console.log("deleteCommentFailure");
               console.log(error);
               toast(
