+++ app/frontend/js/components/delete_draft_button.vue
@@ -1,49 +1,40 @@
 <template>
-  <button :disabled="isToastRunning" @click="deleteComment" class="btn btn-outline-danger">削除</button>
+  <button @click="deleteDraft" type="button" class="btn btn-outline-danger">削除</button>
 </template>
 
 <script>
 import Axios from "axios";
 import { csrfToken } from "@rails/ujs";
 import { async } from "q";
-import Toast from "../mixin/toast";
+import Toast from "../mixins/toast";
 
 Axios.defaults.headers.common["X-CSRF-TOKEN"] = csrfToken();
 export default {
   mixins: [Toast],
   props: {
-    commentId: Number
-  },
-  data: function() {
-    return {};
-  },
-  computed: {
-    isToastRunning: function() {
-      return Swal.isTimerRunning();
-    }
+    articleId: Number
   },
   methods: {
-    deleteComment: function() {
-      const element = this.$el.parentNode;
+    deleteDraft: function() {
+      const element = this.$el.parentNode.parentNode;
       const toast = this.toast;
-      const confirmation = Swal.fire({
+      Swal.fire({
         type: "warning",
-        title: "コメントを削除してもよろしいですか？",
-        text: "この操作は取り消しできません",
+        title: "下書きを削除してもよろしいですか？",
+        text: "この操作は取り消せません",
         showCancelButton: true,
         confirmButtonText: "削除する",
         cancelButtonText: "キャンセル"
       }).then(result => {
         if (result.value) {
-          Axios.delete(`/comments/${this.commentId}`)
-            .then(function(response) {
-              console.log(response.data);
+          Axios.delete(`/drafts/${this.articleId}`)
+            .then(response => {
               if (response.status === 200) {
                 element.remove();
                 toast(response.data.type, response.data.message);
               }
             })
-            .catch(function(error) {
+            .catch(error => {
               console.log("deleteCommentFailure");
               console.log(error);
               toast(
