+++ app/frontend/js/components/favorite_button.vue
@@ -26,7 +26,7 @@
 <script>
 import Axios from "axios";
 import { csrfToken } from "@rails/ujs";
-import Processing from "../mixin/processing";
+import Processing from "../mixins/processing";
 
 Axios.defaults.headers.common["X-CSRF-TOKEN"] = csrfToken();
