+++ app/frontend/js/components/follow_button.vue
@@ -19,7 +19,7 @@
 import Axios from "axios";
 import { csrfToken } from "@rails/ujs";
 import { async } from "q";
-import Processing from "../mixin/processing";
+import Processing from "../mixins/processing";
 
 Axios.defaults.headers.common["X-CSRF-TOKEN"] = csrfToken();
 export default {
