+++ app/frontend/js/components/profile_photo_form.vue
