+++ app/frontend/js/components/remember_button.vue
@@ -13,7 +13,7 @@
 import Axios from "axios";
 import { csrfToken } from "@rails/ujs";
 import { async } from "q";
-import Processing from "../mixin/processing";
+import Processing from "../mixins/processing";
 
 Axios.defaults.headers.common["X-CSRF-TOKEN"] = csrfToken();
 export default {
