+++ app/frontend/scss/_custom_article_show.scss
