+++ app/frontend/scss/_custom_card.scss
