+++ app/frontend/scss/_list-item.scss
