+++ app/frontend/scss/_section.scss
