+++ app/frontend/scss/_separator.scss
