+++ config/webpacker.yml
@@ -2,7 +2,7 @@
 
 default: &default
   source_path: app/frontend
-  source_entry_path: packs
+  source_entry_path: entry
   public_root_path: public
   public_output_path: packs
   cache_path: tmp/cache/webpacker
